More tightening of code
Realized the writing of the PID file no longer needed to be so spread out in
the code, so I could consolidate that.  Once that happened, it's pretty
clear what the code is doing, so removed a bunch of unneeded comments.

Removed nearly 50 lines of code with these past few changes.  Woot!