Code change per Mark's comments
Removed references to div().  Mark commented that the code fragment:

	faclev = div(value,8);

	msg.facility = faclev.quot;
	msg.level    = facvel.rem;

Might not be that great on some dumb compilers, but most all compilers would
convert the div to a shift and the modulus to an AND instruction.  Hey, why
not?  It removes three lines of code, and that's always good.