Bug fix---fix the removal of the firewall rules
Instead of waiting for a period of time per entry, just accumulate them
until we can't, then remove an entry.  This is because there seems to be a
limit to the number of iptable rules, and the entries accumulate faster than
they are removed.  Pity the scammers make it so we can't have nice things on
the Internet.