Bug fix---still fixing the SSH blocking module
  Okay, now we keep the count after adding (when 5, and only 5 attempts have
been seen), and clean up extraneous IP entries at periodic intervals.  The
problem is: when we use '>= 5', then when a bot has multiple attempts at
once finally fail, we get a burst of matches and can end up with multiple
copies of an iptables rule.  If we use '== 5', then we don't get the rules,
but when we clean out the match in the IP list, the others will recreate the
count and potentially add a reduntant rule if there are too many of them,
and even if the don't, the entry won't be cleaned out until the rule is
dropped (after who knows how long) and we might just slowly grow memory when
we don't have to.

  So now we don't clear the IP after adding the iptables rule, but in a
repeating job.